requirements.txt
ScanService/requirements.txt
pyyaml
paho-mqtt
